+++ packages/pangea-sdk/pangea/services/intel.py
@@ -1,5 +1,6 @@
 # Copyright 2022 Pangea Cyber Corporation
 # Author: Pangea Cyber Corporation
+import hashlib
 from typing import Dict, List, Optional
 
 from pangea.response import APIRequestModel, APIResponseModel, PangeaResponse, PangeaResponseResult
@@ -9,7 +10,7 @@ from .base import ServiceBase
 
 class FileLookupRequest(APIRequestModel):
     """
-    TODO: complete
+    File lookup request data
 
     file_hash (str): Hash of the file to be looked up
     hash_type (str): Type of hash, can be "sha256", "sha" or "md5"
@@ -27,7 +28,7 @@ class FileLookupRequest(APIRequestModel):
 
 class FileLookupData(APIResponseModel):
     """
-    TODO: complete
+    File lookup information
     """
 
     category: List[str]
@@ -37,7 +38,7 @@ class FileLookupData(APIResponseModel):
 
 class FileLookupResult(PangeaResponseResult):
     """
-    TODO: complete
+    File lookup result information
     """
 
     data: FileLookupData
@@ -47,7 +48,7 @@ class FileLookupResult(PangeaResponseResult):
 
 class IPLookupRequest(APIRequestModel):
     """
-    TODO: complete
+    IP lookup request data
 
     ip (str): IP address to be looked up
     provider (str, optional): Provider of the reputation information. ("reversinglabs"). Default provider defined by the configuration.
@@ -61,9 +62,29 @@ class IPLookupRequest(APIRequestModel):
     provider: Optional[str] = None
 
 
+class IPLookupData(APIResponseModel):
+    """
+    IP lookup information
+    """
+
+    category: List[str]
+    score: int
+    verdict: str
+
+
+class IPLookupResult(PangeaResponseResult):
+    """
+    IP lookup result
+    """
+
+    data: IPLookupData
+    parameters: Optional[Dict] = None
+    raw_data: Optional[Dict] = None
+
+
 class DomainLookupRequest(APIRequestModel):
     """
-    TODO: complete
+    Domain lookup request data
 
     domain (str): Domain address to be looked up
     provider (str, optional): Provider of the reputation information. ("domaintools"). Default provider defined by the configuration.
@@ -79,7 +100,7 @@ class DomainLookupRequest(APIRequestModel):
 
 class DomainLookupData(APIResponseModel):
     """
-    TODO: complete
+    Domain lookup information
     """
 
     category: List[str]
@@ -89,10 +110,46 @@ class DomainLookupData(APIResponseModel):
 
 class DomainLookupResult(PangeaResponseResult):
     """
-    TODO: complete
+    Domain lookup result
     """
 
     data: DomainLookupData
+    parameters: Optional[Dict] = None
+    raw_data: Optional[Dict] = None
+
+
+class URLLookupRequest(APIRequestModel):
+    """
+    URL lookup request data
+
+    url (str): URL address to be looked up
+    provider (str, optional): Provider of the reputation information. ("crowdstrike"). Default provider defined by the configuration.
+    verbose (bool, optional): Echo back the parameters of the API in the response
+    raw (bool, optional): Return additional details from the provider.
+    """
+
+    url: str
+    verbose: Optional[bool] = None
+    raw: Optional[bool] = None
+    provider: Optional[str] = None
+
+
+class URLLookupData(APIResponseModel):
+    """
+    URL lookup information
+    """
+
+    category: List[str]
+    score: int
+    verdict: str
+
+
+class URLLookupResult(PangeaResponseResult):
+    """
+    URL lookup result
+    """
+
+    data: URLLookupData
     parameters: Optional[Dict] = None
     raw_data: Optional[Dict] = None
 
@@ -138,7 +195,11 @@ class FileIntel(ServiceBase):
         Retrieve file reputation from a provider, using the file's hash.
 
         Args:
-            input (FileLookupInput): input with file information to perform request
+            hash (str): hash to perform lookup
+            hash_type (str): hash type of hash parameter
+            provider (str, optional): intel provider to perfome lookup
+            verbose (bool, optional): true to get more detalied response
+            raw (bool, optional): true to get provider raw response
 
         Raises:
             PangeaAPIException: If an API Error happens
@@ -148,7 +209,7 @@ class FileIntel(ServiceBase):
                 response.result field.  Available response fields can be found in our [API documentation](https://pangea.cloud/docs/api/file-intel).
 
         Examples:
-            response = file_intel.lookup(FileLookupInput(hash="142b638c6a60b60c7f9928da4fb85a5a8e1422a9ffdc9ee49e17e56ccca9cf6e", hash_type="sha256", provider="reversinglabs"))
+            response = file_intel.lookup(hash="142b638c6a60b60c7f9928da4fb85a5a8e1422a9ffdc9ee49e17e56ccca9cf6e", hash_type="sha256", provider="reversinglabs")
 
             \"\"\"
             response contains:
@@ -171,6 +232,63 @@ class FileIntel(ServiceBase):
             \"\"\"
         """
         input = FileLookupRequest(hash=hash, hash_type=hash_type, verbose=verbose, raw=raw, provider=provider)
+        response = self.request.post("lookup", data=input.dict(exclude_none=True))
+        response.result = FileLookupResult(**response.raw_result)
+        return response
+
+    def lookupFilepath(
+        self,
+        filepath: str,
+        provider: Optional[str] = None,
+        verbose: Optional[bool] = None,
+        raw: Optional[bool] = None,
+    ) -> PangeaResponse[FileLookupResult]:
+        """
+        Look up a file
+
+        Retrieve file reputation from a provider, using the file's hash.
+
+        Args:
+            filepath (str): path file to calculate hash and request a lookup
+            provider (str, optional): intel provider to perfome lookup
+            verbose (bool, optional): true to get more detalied response
+            raw (bool, optional): true to get provider raw response
+
+        Raises:
+            PangeaAPIException: If an API Error happens
+
+        Returns:
+            A PangeaResponse where the sanctioned source(s) are in the
+                response.result field.  Available response fields can be found in our [API documentation](https://pangea.cloud/docs/api/file-intel).
+
+        Examples:
+            response = file_intel.lookup(filepath="./myfile.exe", provider="reversinglabs"))
+
+            \"\"\"
+            response contains:
+            {
+                "request_id": "prq_snooq62g4jsolhhpm4ze6pgzhmguflnl",
+                "request_time": "2022-10-10T21:54:19.392Z",
+                "response_time": "2022-10-10T21:54:19.933Z",
+                "status": "Success",
+                "summary": "Hash was found",
+                "result": {
+                    "data": {
+                        "category": [
+                            "Trojan"
+                        ],
+                        "score": 100,
+                        "verdict": "malicious"
+                    }
+                }
+            }
+            \"\"\"
+        """
+
+        data = open(filepath, "rb")
+        hash = hashlib.sha256(data.read()).hexdigest()
+
+        input = FileLookupRequest(hash=hash, hash_type="sha256", verbose=verbose, raw=raw, provider=provider)
         response = self.request.post("lookup", data=input.dict(exclude_none=True))
         response.result = FileLookupResult(**response.raw_result)
         return response
@@ -212,7 +330,10 @@ class DomainIntel(ServiceBase):
         Retrieve Domain reputation from a provider.
 
         Args:
-            input (URLLookupInput): input with domain information to perform request
+            domain (str): domain to request for a lookup
+            provider (str, optional): intel provider to perfome lookup
+            verbose (bool, optional): true to get more detalied response
+            raw (bool, optional): true to get provider raw response
 
         Raises:
             PangeaAPIException: If an API Error happens
@@ -222,7 +343,7 @@ class DomainIntel(ServiceBase):
                 response.result field.  Available response fields can be found in our [API documentation](https://pangea.cloud/docs/api/domain-intel).
 
         Examples:
-            response = domain_intel.lookup(DomainLookupInput(domain="737updatesboeing.com", provider="domaintools"))
+            response = domain_intel.lookup(domain="737updatesboeing.com", provider="domaintools")
 
             \"\"\"
             response contains:
@@ -252,4 +373,149 @@ class DomainIntel(ServiceBase):
         input = DomainLookupRequest(domain=domain, verbose=verbose, provider=provider, raw=raw)
         response = self.request.post("lookup", data=input.dict(exclude_none=True))
         response.result = DomainLookupResult(**response.raw_result)
+        return response
+
+
+class IpIntel(ServiceBase):
+    """IP Intel service client
+
+    Provides methods to interact with [Pangea IP Intel Service](/docs/api/ip-intel)
+
+    The following information is needed:
+        PANGEA_TOKEN - service token which can be found on the Pangea User
+            Console at [https://console.pangea.cloud/project/tokens](https://console.pangea.cloud/project/tokens)
+
+    Examples:
+        import os
+
+        # Pangea SDK
+        from pangea.config import PangeaConfig
+        from pangea.services import IpIntel
+
+        PANGEA_TOKEN = os.getenv("PANGEA_TOKEN")
+
+        ip_intel_config = PangeaConfig(domain="pangea.cloud")
+
+        # Setup Pangea IP Intel service
+        ip_intel = IpIntel(token=PANGEA_TOKEN, config=ip_intel_config)
+    """
+
+    service_name = "ip-intel"
+    version = "v1"
+
+    def lookup(
+        self, ip: str, verbose: Optional[bool] = None, raw: Optional[bool] = None, provider: Optional[str] = None
+    ) -> PangeaResponse[IPLookupResult]:
+        """
+        Retrieve IP address reputation from a provider.
+
+        Args:
+            input (IPLookupInput): input with IP information to perform request
+
+        Raises:
+            PangeaAPIException: If an API Error happens
+
+        Returns:
+            A PangeaResponse where the sanctioned source(s) are in the
+                response.result field.  Available response fields can be found in our [API documentation](/docs/api/ip-intel)
+
+        Examples:
+            response = ip_intel.lookup(IPLookupInput(ip="93.231.182.110", provider="crowdstrike"))
+
+            \"\"\"
+            response contains:
+            {
+                "request_id": "prq_xoohakngaerteg4yiekikva3issxp4bq",
+                "request_time": "2022-08-23T03:28:20.225Z",
+                "response_time": "2022-08-23T03:28:20.244Z",
+                "status": "success",
+                "summary": "IP was found",
+                "result": {
+                    "data": {
+                        "category": [
+                            "Suspicious"
+                        ],
+                        "score": 0,
+                        "verdict": "malicious"
+                    }
+                }
+            }
+            \"\"\"
+        """
+        input = IPLookupRequest(ip=ip, verbose=verbose, raw=raw, provider=provider)
+        response = self.request.post("lookup", data=input.dict(exclude_none=True))
+        response.result = IPLookupResult(**response.raw_result)
+        return response
+
+
+class UrlIntel(ServiceBase):
+    """URL Intel service client.
+
+    Provides methods to interact with [Pangea URL Intel Service](/docs/api/url-intel)
+
+    The following information is needed:
+        PANGEA_TOKEN - service token which can be found on the Pangea User
+            Console at [https://console.pangea.cloud/project/tokens](https://console.pangea.cloud/project/tokens)
+
+    Examples:
+        import os
+
+        # Pangea SDK
+        from pangea.config import PangeaConfig
+        from pangea.services import UrlIntel
+
+        PANGEA_TOKEN = os.getenv("PANGEA_TOKEN")
+
+        url_intel_config = PangeaConfig(domain="pangea.cloud")
+
+        # Setup Pangea URL Intel service
+        url_intel = UrlIntel(token=PANGEA_TOKEN, config=url_intel_config)
+    """
+
+    service_name = "url-intel"
+    version = "v1"
+
+    def lookup(
+        self, url: str, verbose: Optional[bool] = None, raw: Optional[bool] = None, provider: Optional[str] = None
+    ) -> PangeaResponse[URLLookupResult]:
+        """
+        Retrieve URL address reputation from a provider.
+
+        Args:
+            input (URLLookupInput): input with URL information to perform request
+
+        Raises:
+            PangeaAPIException: If an API Error happens
+
+        Returns:
+            A PangeaResponse where the sanctioned source(s) are in the
+                response.result field.  Available response fields can be found in our [API documentation](/docs/api/url-intel)
+
+        Examples:
+            response = url_intel.lookup(URLLookupInput(url="http://113.235.101.11:54384", provider="crowdstrike"))
+
+            \"\"\"
+            response contains:
+            {
+                "request_id": "prq_5ugxruda7vmsgioup6vjvaqmnmvxzbqv",
+                "request_time": "2022-08-23T03:40:03.549Z",
+                "response_time": "2022-08-23T03:40:03.694Z",
+                "status": "success",
+                "summary": "Url was found",
+                "result": {
+                    "data": {
+                        "category": [
+                            "Not Provided"
+                        ],
+                        "score": 80,
+                        "verdict": "malicious"
+                    }
+                }
+            }
+            \"\"\"
+        """
+
+        input = URLLookupRequest(url=url, provider=provider, verbose=verbose, raw=raw)
+        response = self.request.post("lookup", data=input.dict(exclude_none=True))
+        response.result = URLLookupResult(**response.raw_result)
         return response