+++ CHANGELOG.md
@@ -5,12 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2023-01-05
+
+### Added
+
+- Intel add IP and URL services with lookup endpoint
+
+
+## [1.0.2] - 2022-12-23
+
+### Fixed
+
+- Multiples bugs in audit tools used to dump and verify events
+
+
+## [1.0.1] - 2022-12-19
 
 ### Added
 
 - This CHANGELOG
 - Test to search with dates as filter
+- Functions to get token and domain according to test environment (PROD/DEV/STG)
 
 ### Fixed
 
@@ -40,5 +55,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact client
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-python/compare/v1.0.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-python/compare/v1.1.0...main
+[1.1.0]: https://github.com/pangeacyber/pangea-python/compare/v1.0.2...v1.1.0
+[1.0.2]: https://github.com/pangeacyber/pangea-python/compare/v1.0.1...v1.0.2
+[1.0.1]: https://github.com/pangeacyber/pangea-python/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/pangeacyber/pangea-python/releases/tag/v1.0.0